# manifest: avoid BlobReferences interface allocation